Code restructure complete
Fully separated 42 communication code from flight software. Added
defines to disable UART and SPI in the state machine. The user now needs
to set UART and SPI handles manually before running the ACS.

+++ Mech_Lib/PWM_Library.c
@@ -1,6 +1,6 @@
 #include "PWM_Library.h"
 
-void PWM_Set_Duty_Cycle(TIM_HandleTypeDef *htim, uint16_t duty_cycle, uint32_t channel)
+void PWM_Set_Duty_Cycle(TIM_HandleTypeDef *htim, float duty_cycle, uint32_t channel)
 {
 	switch(channel) {
 		case TIM_CHANNEL_1:
